CD_016_database_replication/app/requirements.txt
# requirements.txt
psycopg2-binary>=2.9
rich>=12